chore(fix):
- refactor routes to use one instance of express.Router()
