+++ server/api/v1/routes/index.js
@@ -1,14 +1,7 @@
-import express from 'express';
-import authRouter from './auth';
+import authRoutes from './auth';
 
-const router = express.Router();
+const routes = (router) => {
+  authRoutes(router);
+};
 
-router.get('/', (req, res) => {
-  res.status(200).send({
-    message: 'Welcome to Politico API',
-    docs: 'https://politicobooth.docs.apiary.io/',
-  });
-});
-router.use(authRouter);
-
-export default router;
+export default routes;
